-rename dateFomatter to dates -add countdown time dependency -add basic stub for countdown timer
