--- src/components/Footer/CountdownTimer.jsx
@@ -1,6 +1,0 @@
-const CountdownTimer = () => {
-
-    return(<p>OO:00:00</p>)
-}
- 
-export default CountdownTimer;
